requirements.txt
beautifulsoup4==4.11.1
pandas==1.3.4
path.py==12.5.0
requests==2.26.0
tqdm==4.62.3
lxml==4.9.1